% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download a kaggle dataset}
\usage{
get_data(datasets, check_size = TRUE)
}
\arguments{
\item{datasets}{either a dataset returned from the function \code{search_kaggle}
or a (vector) of (character) references of the form \if{html}{\out{<kaggle user>}}/\if{html}{\out{<data set>}}.}

\item{check_size}{boolean, check that the size is not in GB}
}
\value{
tibble of the input expanded by a list variable \code{data}
}
\description{
Download a kaggle dataset
}
