% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_search_page.R
\name{get_search_page}
\alias{get_search_page}
\title{Helper function: get page X from the search results}
\usage{
get_search_page(search_str, page)
}
\arguments{
\item{search_str}{character value, key word of the search}

\item{page}{integer value, page number}
}
\value{
data.frame
}
\description{
This function accesses one page of the returned search results, downloads the information,
and returns a dataset of this page's results.
The user should not have to call this function directly in interactive use.
}
