% Generated by roxygen2: do not edit by hand
% Please edit documentation in make-the-kaggling-package.Rmd.
\docType{data}
\name{kaggle_diamonds}
\alias{kaggle_diamonds}
\title{A start on a description}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{x}{a predictor variable}
\item{y}{a response variable}
}
}
\source{
This data was simulated.  But usually we might cite a source here.
}
\usage{
kaggle_diamonds
}
\description{
A start on a description
}
\keyword{datasets}