% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggling-package.R
\docType{package}
\name{kaggling-package}
\alias{kaggling}
\alias{kaggling-package}
\title{kaggling: Searches Kaggle Platform for Data by Keywords}
\description{
Kaggle is used as a platform to offer and participate in data competitions by thousands of users. The platform features an easy upload mechanism for data sets. To date almost 300k data sets are available at the platform. This package provides access to a key-word based search of available data sets based on the kaggle API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heike/kaggling}
  \item Report bugs at \url{https://github.com/heike/kaggling/issues}
}

}
\author{
\strong{Maintainer}: Heike Hofmann \email{hofmann@iastate.edu} (\href{https://orcid.org/0000-0001-6216-5183}{ORCID})

Authors:
\itemize{
  \item Susan Vanderplas \email{srvander@iastate.edu} (\href{https://orcid.org/0000-0002-3803-0972}{ORCID})
}

}
\keyword{internal}
