% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_kaggle.R
\name{search_kaggle}
\alias{search_kaggle}
\title{Keyword search kaggle datasets}
\usage{
search_kaggle(search_str)
}
\arguments{
\item{search_str}{character value, key word of the search}
}
\value{
data.frame
}
\description{
This function searches the kaggle platform for all datasets fitting the keyword search <search_str>
and returns available meta information.
}
