% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_byte.R
\name{size_MB}
\alias{size_MB}
\title{Convert size character vector to a size in byte}
\usage{
size_MB(col)
}
\arguments{
\item{col}{vector of characters with size information}
}
\value{
numeric vector of the same length as col with the size information in megabyte.
}
\description{
Size should really be an S3 object but here's my compromise:
convert the humanly-readable character string to something we can
work with analytically.
}
