% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Download meta information for a kaggle dataset}
\usage{
get_meta(datasets)
}
\arguments{
\item{datasets}{either a dataset returned from the function \code{search_kaggle}
or a (vector) of (character) references of the form \if{html}{\out{<kaggle user>}}/\if{html}{\out{<data set>}}.}
}
\value{
tibble of the input expanded by a list column \code{meta}
}
\description{
Download meta information for a kaggle dataset
}
