% Generated by roxygen2: do not edit by hand
% Please edit documentation in make-the-kaggling-package.Rmd.
\docType{package}
\name{kaggling-package}
\alias{kaggling}
\alias{kaggling-package}
\title{kaggling: Searches kaggles for data by keywords}
\description{
Kaggle is used as a platform to offer and participate in data competitions by thousands of users. The platform features an easy upload mechanism for data sets. To date almost 300k data sets are available at the platform. This package provides access to a key-word based search of available data sets based on the kaggle API.
}
\author{
\strong{Maintainer}: Heike Hofmann \email{hofmann@iastate.edu}

}
\keyword{internal}